+++ CHANGELOG.md
@@ -52,6 +52,7 @@ We used to have incremental versioning before `0.1.0`.
 - Improves violation code testing
 - Improves testing of `.. versionchanged` and `previous_codes` properties
 - Reference `isort` settings requirement for compliance with `WSP318` in docstring
+- Forbids too long call chains like `foo(a)(b)(c)(d)`
 
 
 ## 0.12.5